auth redirects & flash msg design

+++ controllers/postController.js
@@ -15,10 +15,16 @@ module.exports = {
             });
     },
     showView: (req, res) => {
-        res.render("post/post.ejs")
+        res.render("post/post.ejs");
     },
-    new: (req, res) => {
-        res.render("post/create_post.ejs")
+    new: (req, res, next) => {
+        if(req.isAuthenticated()) {
+            res.render("post/create_post.ejs")
+        }else{
+            res.locals.redirect = "/";
+            req.flash("error", `You are not logged in`);
+            next();
+        }
     },
     create: (req, res, next) => {
         let postParams = {
@@ -58,14 +64,26 @@ module.exports = {
     },
     editView: (req, res, next) => {
         let postId = req.params.postId;
-        Post.findById(postId).exec()
-            .then(post => {
-                res.render("post/edit_post.ejs", {post: post});
-            })
-            .catch(error => {
-                console.log(`Error fetching post by ID: ${postId}\n${error.message}`);
-                next(error);
-            });
+        if(req.isAuthenticated()) {
+            Post.findById(postId).populate("user").exec()
+                .then(post => {
+                    if(req.user.isAdmin || (post.user === req.user)) {
+                        res.render("post/edit_post.ejs", {post: post});
+                    }else{
+                        res.locals.redirect = "/post/" + postId;
+                        req.flash("error", `You are not authorised to edit this post`);
+                        next();
+                    }
+                })
+                .catch(error => {
+                    console.log(`Error fetching post by ID: ${postId}\n${error.message}`);
+                    next(error);
+                });
+        }else{
+            res.locals.redirect = "/post/" + postId;
+            req.flash("error", `You are not logged in`);
+            next();
+        }
     },
     update: (req, res, next) => {
         let postId = req.params.postId;
@@ -96,38 +114,50 @@ module.exports = {
     },
     delete: (req, res, next) => {
         let postId = req.params.postId;
-        Post.findById(postId).exec()
-            .then(post => {
-                User.findByIdAndUpdate(post.user, {
-                    $pull: {posts: postId}
-                }).exec()
-                    .then(() => {
-                        Post.findByIdAndRemove(postId).exec()
+        if(req.isAuthenticated()) {
+            Post.findById(postId).exec()
+                .then(post => {
+                    if((req.user.isAdmin) || (post.user === req.user)) {
+                        User.findByIdAndUpdate(post.user, {
+                            $pull: {posts: postId}
+                        }).exec()
                             .then(() => {
-                                res.locals.redirect = "/";
-                                req.flash("success", `Post ${postId} deleted successfully!`);
-                                console.log(`Deleted Post: ${postId}`);
-                                next();
+                                Post.findByIdAndRemove(postId).exec()
+                                    .then(() => {
+                                        res.locals.redirect = "/";
+                                        req.flash("success", `Post ${postId} deleted successfully!`);
+                                        console.log(`Deleted Post: ${postId}`);
+                                        next();
+                                    })
+                                    .catch(error => {
+                                        res.locals.redirect = `/post/${postId}`;
+                                        req.flash("error", `Failed to delete post ${postId} because: ${error.message}`);
+                                        console.log(`Error deleting post by ID: ${postId}\n${error.message}`);
+                                        next();
+                                    });
                             })
                             .catch(error => {
                                 res.locals.redirect = `/post/${postId}`;
-                                req.flash("error", `Failed to delete post ${postId} because: ${error.message}`);
-                                console.log(`Error deleting post by ID: ${postId}\n${error.message}`);
+                                req.flash("error", `Failed to remove post ${postId} from User because: ${error.message}`);
+                                console.log(`Error removing post by User: ${post.user}\n${error.message}`);
                                 next();
                             });
-                    })
-                    .catch(error => {
-                        res.locals.redirect = `/post/${postId}`;
-                        req.flash("error", `Failed to remove post ${postId} from User because: ${error.message}`);
-                        console.log(`Error removing post by User: ${post.user}\n${error.message}`);
+                    } else {
+                        res.locals.redirect = "/post/" + postId;
+                        req.flash("error", `You are not authorised to delete this post`);
                         next();
-                    });
-            })
-            .catch(error => {
-                res.locals.redirect = `/post/${postId}`;
-                req.flash("error", `Cannot find User ${post.user} because: ${error.message}`);
-                console.log(`Error finding User ${post.user}: ${error.message}`)
-            })
+                    }
+                })
+                .catch(error => {
+                    res.locals.redirect = `/post/${postId}`;
+                    req.flash("error", `Cannot find User ${post.user} because: ${error.message}`);
+                    console.log(`Error finding User ${post.user}: ${error.message}`)
+                })
+        } else {
+            res.locals.redirect = "/post/" + postId;
+            req.flash("error", `You are not logged in`);
+            next();
+        }
     },
     redirectView: (req, res, next) => {
         let redirectPath = res.locals.redirect;
